hf_deployment/requirements.txt
# Face Recognition System - Dependencies
# Python 3.8+

# ============================================================================
# IMPORTANT: Deep Learning Frameworks dengan CUDA Support
# ============================================================================
# 
# OPTION 1 - CUDA 11.8 (Recommended untuk RTX 30xx/40xx series):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
#
# OPTION 2 - CUDA 12.1 (Untuk GPU terbaru):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#
# OPTION 3 - CPU Only (jika tidak ada GPU):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
#
# NOTE: Install PyTorch FIRST sebelum requirements.txt lainnya
# ============================================================================

# Deep Learning Frameworks
# torch>=2.0.0  # COMMENT: Install manually dengan command di atas
# torchvision>=0.15.0  # COMMENT: Install manual bersama torch
tensorflow>=2.13.0
keras>=2.13.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
mtcnn>=0.1.1
facenet-pytorch>=2.5.3

# Image Processing
Pillow>=10.0.0
pillow-heif>=0.13.0  # For HEIC conversion
albumentations>=1.3.1

# Deep Learning Models
timm>=0.9.0  # PyTorch Image Models (for ViT/DeiT)
transformers>=4.30.0  # HuggingFace Transformers

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Desktop Application
PyQt5>=5.15.9
gradio>=3.50.0

# Utilities
pyyaml>=6.0
colorlog>=6.7.0
python-dotenv>=1.0.0
